refactor: CustomAdapterを別ファイルに分割
diffUtilはそのままCustomAdapterに埋め込んだ
